+++ spec/directions/bake_numbered_table/v1_spec.rb
@@ -60,6 +60,17 @@ RSpec.describe Kitchen::Directions::BakeNumberedTable::V1 do
     ).tables.first
   end
 
+  let(:unstyled_table) do
+    book_containing(html:
+      one_chapter_with_one_page_containing(
+        <<~HTML
+          <table class="unstyled" id="tId">
+          </table>
+        HTML
+      )
+    ).tables.first
+  end
+
   let(:other_table) do
     book_containing(html:
       one_chapter_with_one_page_containing(
@@ -167,6 +178,25 @@ RSpec.describe Kitchen::Directions::BakeNumberedTable::V1 do
       <<~HTML
         <div class="os-table os-text-heavy-container">
           <table class="text-heavy" id="tId">
+        </table>
+          <div class="os-caption-container">
+            <span class="os-title-label">Table </span>
+            <span class="os-number">2.3</span>
+            <span class="os-divider"> </span>
+            <span class="os-divider"> </span>
+          </div>
+        </div>
+      HTML
+    )
+  end
+
+  it 'bakes an unstyled table' do
+    described_class.new.bake(table: unstyled_table, number: '2.3')
+
+    expect(unstyled_table.document.search('.os-table').first).to match_normalized_html(
+      <<~HTML
+        <div class="os-table os-unstyled-container">
+          <table class="unstyled" id="tId">
         </table>
           <div class="os-caption-container">
             <span class="os-title-label">Table </span>